✨ Split exception handling per arch + Move GcInfo to Garbage Collector
